+++ candy-machine/program/src/processor/collection/mod.rs
@@ -1,0 +1,7 @@
+pub mod remove_collection;
+pub mod set_collection;
+pub mod set_collection_during_mint;
+
+pub use remove_collection::*;
+pub use set_collection::*;
+pub use set_collection_during_mint::*;
