+++ candy-machine/program/src/processor/mod.rs
@@ -1,0 +1,13 @@
+pub mod add_config_lines;
+pub mod collection;
+pub mod initialize;
+pub mod mint;
+pub mod update;
+pub mod withdraw;
+
+pub use add_config_lines::*;
+pub use collection::*;
+pub use initialize::*;
+pub use mint::*;
+pub use update::*;
+pub use withdraw::*;
